Major refactoring to make BIRE more generally applicable
Documents have a Prior object which is used to initialize a sampling
chain given a document.
AnnotatedDocuments provide a result object that is given to the
objective function alongside the state to score the given state.
Initializer interface that is used to initialize the sampling chain
given a document with a specific prior object (e.g. an initial state)
JoinExplorer merges the results of individual explorers into a single
list.